+++ lib/overcommit/hook/base.rb
@@ -17,6 +17,18 @@ module Overcommit::Hook
       raise NotImplementedError, 'Hook must define `run`'
     end
 
+    # Runs the hook and transforms the status returned based on the hook's
+    # configuration.
+    #
+    # Poorly named because we already have a bunch of hooks in the wild that
+    # implement `#run`, and we needed a wrapper step to transform the status
+    # based on any custom configuration.
+    def run_and_transform
+      status, output = run
+
+      [transform_status(status), output]
+    end
+
     def name
       self.class.name.split('::').last
     end
@@ -92,6 +104,20 @@ module Overcommit::Hook
                     File::FNM_PATHNAME | # Wildcard doesn't match separator
                     File::FNM_DOTMATCH   # Wildcards match dotfiles
       )
+    end
+
+    # Transforms the hook's status based on custom configuration.
+    #
+    # This allows users to change failures into warnings, or vice versa.
+    def transform_status(status)
+      case status
+      when :fail, :bad
+        @config.fetch('on_fail', :fail).to_sym
+      when :warn
+        @config.fetch('on_warn', :warn).to_sym
+      else
+        status
+      end
     end
   end
 end